+++ examples/pong.rs
@@ -4,14 +4,31 @@ use std::rc::Rc;
 use std::thread;
 use std::time::Duration;
 
+#[repr(i32)]
+#[derive(Copy, Clone)]
+enum Direction {
+    Positive = 1,
+    Negative = -1,
+}
+
+struct Ball {
+    wid: valuator::FillDial,
+    pos: (i32, i32),             // x and y positions
+    dir: (Direction, Direction), // x and y directions
+}
+
 fn main() {
     let app = app::App::default();
     let mut wind = window::Window::default()
         .with_size(800, 600)
         .center_screen()
         .with_label("Pong!");
-    let mut ball = valuator::FillDial::new(0, 0, 40, 40, "");
-    ball.set_color(Color::White);
+    let mut ball = Ball {
+        wid: valuator::FillDial::new(0, 0, 40, 40, ""),
+        pos: (0, 0),
+        dir: (Direction::Positive, Direction::Positive),
+    };
+    ball.wid.set_color(Color::White);
     wind.set_color(Color::Black);
     wind.end();
     wind.show();
@@ -23,57 +40,39 @@ fn main() {
         draw::draw_rectf(*paddle_c.borrow(), 540, 160, 20);
     }));
 
-    let ball_x = &mut 0;
-    let ball_y = &mut 0;
     let paddle_c = paddle_pos.clone();
-    let mut wind_c = wind.clone();
-
     wind.handle(Box::new(move |ev| {
         match ev {
-            enums::Event::KeyDown => {
-                match app::event_key() {
-                    enums::Key::Right => *paddle_pos.borrow_mut() += 30,
-                    enums::Key::Left => *paddle_pos.borrow_mut() -= 30,
-                    _ => (),
-                }
-                wind_c.redraw();
+            enums::Event::Move => {
+                *paddle_pos.borrow_mut() = app::event_coords().0 - 80; // Mouse's x position
                 true
             }
             _ => false,
         }
     }));
 
-    let x_positive = &mut true;
-    let y_positive = &mut true;
     while app.wait().unwrap() {
-        if *x_positive {
-            *ball_x += 5;
-        } else {
-            *ball_x -= 5;
-        }
-        if *y_positive {
-            *ball_y += 5;
-        } else {
-            *ball_y -= 5;
-        }
-        if *ball_y == 540 - 40
-            && (*ball_x > *paddle_c.borrow() - 40 && *ball_x < *paddle_c.borrow() + 160)
+        ball.pos.0 += 10 * ball.dir.0 as i32;
+        ball.pos.1 += 10 * ball.dir.1 as i32;
+        if ball.pos.1 == 540 - 40
+            && (ball.pos.0 > *paddle_c.borrow() - 40 && ball.pos.0 < *paddle_c.borrow() + 160)
         {
-            *y_positive = false;
+            ball.dir.1 = Direction::Negative;
         }
-        if *ball_y == 0 {
-            *y_positive = true;
+        if ball.pos.1 == 0 {
+            ball.dir.1 = Direction::Positive;
         }
-        if *ball_x == 800 - 40 {
-            *x_positive = false;
+        if ball.pos.0 == 800 - 40 {
+            ball.dir.0 = Direction::Negative;
         }
-        if *ball_x == 0 {
-            *x_positive = true;
+        if ball.pos.0 == 0 {
+            ball.dir.0 = Direction::Positive;
         }
-        if *ball_y > 600 {
-            app.quit();
+        if ball.pos.1 > 600 {
+            ball.pos = (0, 0);
+            ball.dir = (Direction::Positive, Direction::Positive);
         }
-        ball.resize(*ball_x, *ball_y, 40, 40);
+        ball.wid.resize(ball.pos.0, ball.pos.1, 40, 40);
         wind.redraw();
         thread::sleep(Duration::from_millis(16));
     }